requirements.txt
fastapi[all]
pymongo
pandas
python-dotenv
transformers
scikit-learn
torch
gensim
